+++ NEWS.md
@@ -1,4 +1,4 @@
-# mnirs 0.4.2
+# mnirs 0.4.3
 
 * Create `analyse_kinetics()` to process mNIRS kinetics on nested `nirs_channels` within a list of data.frames.
     * Currently implemented for `method = "peak_slope"` only, with other methods coming soon.
@@ -6,6 +6,13 @@
 * Remove `{tidyr}` dependency from `plot.mnirs()`, and from `{mnirs}` package dependencies.
 * Update `signif_trailing()` to avoid overprinting digits. Used internally for display.
     * `format = c("max_digits", "max_signif")` will print the lesser of either `digits`, or the maximum decimals/sigfigs in the data.
+
+# mnirs 0.4.2
+
+* Update `read_mnirs()` 
+    * Fix internal `read_file()` to better recognise .csv formats where the header has fewer columns than the data table.
+    * If `nirs_channels` is left blank (now default), `read_mnirs()` will attempt to recognise the NIRS device file format by searching for known data table column headers, and return the entire data table as if `keep_all = TRUE`, as a data exploration option.
+    * Update `add_timestamp` to recognise date-time (POSIXct) values in the file header or `time_channel` and add `timestamp` column, and metadata `attr(df, "start_timestamp")`.
 
 # mnirs 0.4.1